% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_error.R
\name{standard_error}
\alias{standard_error}
\title{Calculate the standard error of a sample}
\usage{
standard_error(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{

}
\description{
Calculate the standard error of a sample
}
